requirements.txt
anthropic>=0.18.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
tenacity>=8.2.0
structlog>=23.0.0
rich>=13.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.1.0